Updated team API points

+++ API/src/main/java/org/setup/Listify/Service/TeamsService.java
@@ -1,8 +1,8 @@
 package org.setup.Listify.Service;
 
-import org.setup.Listify.Exception.ListNotFoundException;
-import org.setup.Listify.Exception.TeamMemberNotFoundException;
-import org.setup.Listify.Exception.TeamNotFoundException;
+import jakarta.persistence.EntityManager;
+import jakarta.transaction.Transactional;
+import org.setup.Listify.Exception.*;
 import org.setup.Listify.Model.TeamMembers;
 import org.setup.Listify.Model.TeamProjects;
 import org.setup.Listify.Model.Teams;
@@ -16,6 +16,8 @@ import java.util.List;
 @Service
 public class TeamsService {
     @Autowired
+    private EntityManager entityManager;
+    @Autowired
     private TeamsRepository teamsRepository;
     @Autowired
     private TeamMembersRepository teamMembersRepository;
@@ -26,7 +28,6 @@ public class TeamsService {
         if (teams.isEmpty()) {
             throw new ListNotFoundException("teams");
         }
-
         return teams;
     }
 
@@ -50,6 +51,15 @@ public class TeamsService {
                 .orElseThrow(() -> new TeamMemberNotFoundException(userID));
     }
 
+    public TeamMembers getTeamMemberByUserId(Long userID) {
+        List<TeamMembers> teamMembersList = teamMembersRepository.findByUserID(userID);
+
+        if (teamMembersList.isEmpty()) {
+            throw new UserNotFoundException("User with id: " + userID + " not found");
+        }
+        return teamMembersList.getFirst();
+    }
+
     public List<TeamMembers> getTeamMembersByTeamId(Long teamID) {
         getTeamById(teamID);
         List<TeamMembers> teamMembers = teamMembersRepository.findByTeamID(teamID);
@@ -61,11 +71,20 @@ public class TeamsService {
         return teamMembers;
     }
 
+    public boolean isTeamLeader(Long teamID, Long userID) {
+        boolean isTeamLeader = teamMembersRepository.existsByTeamIDAndUserIDAndIsTeamLeaderTrue(teamID, userID);
+        if (!isTeamLeader) {
+            throw new NotTeamLeaderException("User with ID " + userID + " is not the team leader for team ID " + teamID);
+        }
+        return isTeamLeader;
+    }
+
+
     public List<TeamProjects> getProjectsByTeamID(Long teamID) {
         getTeamById(teamID);
         List<TeamProjects> teamProjects = teamsRepository.findProjectsByTeamID(teamID);
 
-        if(teamProjects.isEmpty()) {
+        if (teamProjects.isEmpty()) {
             throw new ListNotFoundException("team projects for team " + teamID);
         }
 
@@ -73,7 +92,8 @@ public class TeamsService {
     }
 
     public Long addTeam(Long userID, String teamName) {
-        getByTeamMembersId(userID);
+        TeamMembers teamMember = getTeamMemberByUserId(userID);
+
         teamsRepository.createTeam(userID, teamName);
 
         Teams lastCreatedTeam = teamsRepository.findTopByOrderByTeamIDDesc();
@@ -82,27 +102,51 @@ public class TeamsService {
 
     public void assignMemberToTeam(Long teamLeaderID, Long userID, Long teamID) {
         getTeamById(teamID);
+        getTeamMemberByUserId(userID);
+        isTeamLeader(teamID, teamLeaderID);
+        boolean memberExists = teamMembersRepository.existsByTeamIDAndUserID(teamID, userID);
+
+        if (memberExists) {
+            throw new TeamMemberAlreadyExistsException("User with ID " + userID + " is already a member of team " + teamID);
+        }
+
         teamMembersRepository.assignMemberToTeam(teamLeaderID, userID, teamID);
     }
 
     public void deleteTeam(Long teamID, Long teamLeaderID) {
         getTeamById(teamID);
+        isTeamLeader(teamID, teamLeaderID);
         teamsRepository.deleteTeam(teamID, teamLeaderID);
     }
 
     public void deleteMemberFromTeam(Long userID, Long teamID, Long teamLeaderID) {
         getTeamById(teamID);
+        isTeamLeader(teamID, teamLeaderID);
+        boolean memberExists = teamMembersRepository.existsByTeamIDAndUserID(teamID, userID);
+
+        if (!memberExists) {
+            throw new TeamMemberAlreadyExistsException("User with ID " + userID + " is not a member of team " + teamID);
+        }
         teamMembersRepository.deleteMemberFromTeam(userID, teamID, teamLeaderID);
     }
 
-    public void updateTeamDetails(Long teamID, Long teamLeaderID, String newTeamName) {
+    @Transactional
+    public void updateTeamDetails(Long userID, Long teamID, String newTeamName) {
         getTeamById(teamID);
-        teamsRepository.updateTeamDetails(teamID, teamLeaderID, newTeamName);
+        isTeamLeader(teamID, userID);
+        teamsRepository.updateTeamDetails(userID, teamID, newTeamName);
+        entityManager.flush();
+        entityManager.clear();
     }
 
     public void updateTeamLeader(Long teamLeaderID, Long teamID, Long newTeamLeaderID) {
         getTeamById(teamID);
+        isTeamLeader(teamID, teamLeaderID);
+        boolean memberExists = teamMembersRepository.existsByTeamIDAndUserID(teamID, newTeamLeaderID);
+
+        if (!memberExists) {
+            throw new TeamMemberAlreadyExistsException("User with ID " + newTeamLeaderID + " is not a member of team " + teamID);
+        }
         teamMembersRepository.updateTeamLeader(teamLeaderID, teamID, newTeamLeaderID);
     }
-
 }
